requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
pydantic==2.10.6
supabase==2.13.0
python-dotenv==1.0.1